Reduce allocations in TextDocumentStates.AddRange (#75640)
This method was calling ImmutableList.AddRange giving it the result of a Select linq expression. The underlying ImmutableList code doesn't end up special casing that collection type, and thus creates a FallbackWrapper around the enumeration. The FallbackWrapper.Count method uses Enumerable.ToArray to cache a collection object, which ends up

1) doing the double resize allocation dance pass until done
2) resizing that final array

This essentially means we've allocated around 3x the amount of memory needed to hold our documentids. Instead, if we use a pooled list, we can generally avoid the allocations completely (the ImmutableList code instead creates a ListOfTWrapper around our List which doesn't need to allocate)

This method is showing up as around 0.8% of allocations in the typing section of the csharp editing speedometer test.