Reduce allocations in VirtualCharService.CreateVirtualCharSequence (#75654)
This method shows up in OOP allocations in the scrolling speedometer test. This method is invoked to create runes for text sequences for strings. The standard case is that there aren't any multi-char runes in a string, in which case AbstractVirtualCharService.CreateVirtualCharSequence doesn't end up using the ImmutableSegmentedList in the result it creates. This means that we can reuse this builder and it's underlying array after it's been cleared out.

This shows up as around 2.0% of allocations in out OOP in the scrolling speedometer test.