Update breaking change docs
Offline, we had a discussion that the title and H1 were verbose and could be misunderstood.

I've updated the most recent 4 documents to match that discussion.

Notes for reviewers:

All edits in each file should be the same (modulo typos and version numbers). Let's agree on the language on one of them, then make sure they all match.