Modify segmented list to grow by growth rate. (#75756)
By using a growth rate, we can control how often the outer array grows (each growth involves not only reallocating the outer array, but as SegmentedArray doesn't allow nulls, it requires alllocating each inner array). So, by controlling this growth rate, we balance the allocation costs of the outer and inner arrays. Benchmarking (not checked in) demonstrated a growth factor of 12.5% as a good rate for arrays containing up to 64K segments.