Change tagging to free a SegmentedList back to a pool
The GetTags calls into our taggers require passing back an IEnumerable. Previously, we created a SegmentedList to hold all the data and then passed that back. However, that suffers from not being able to add back the returned list into a pool. Instead, switch to using yield return over the populated segmented list as that will allow the list to be freed back to the pool upon completion.

Utilizing the yield state machinery does incur an allocation, but this should be a much smaller allocation than the potentially very large arrays that our taggers were returning.