Reduce some allocations from remote classification serialization (#75911)
SerializableClassifiedSpans.Dehydrate was previously taking in an ImmutableArray, causing several callers to allocate. Instead, have it take in a SegmentedList, which several of the callers already have, and if not, a pooled instance can be obtained and populated.

ClassifiedSpan[] shows up as 5.4% of allocations in our CodeAnalysis process in the platform's scrolling speedometer test.