Reduce formatting allocation in CodeAnalysis process in scrolling speedometer (#75912)
TokenPairWithOperations[] accounts for 2.4% of total allocations in the test. Locally reproducing this demonstrated that there is usually a small number of tokens that are added, but occasionally a much larger token count is added. Doing a single shot allocation in the case where we are significantly increasing the SegmentedList size is more efficient than growing organically by adding single items.