Ensure NFW gets reported before result is reported to client and remove async listener in queue (#75907)
Should resolve
https://dev.azure.com/devdiv/DevDiv/_workitems/edit/2224584

Optprof web application tests have been failing for a while in VS. Tim
tracked down the issue to an async listener introduced in
https://github.com/dotnet/roslyn/pull/74530/

The never completing listener were requests for workspace diagnostics.
This is actually expected for workspace diagnostics requests, as they
are held open indefinitely if nothing changes (see
https://github.com/dotnet/roslyn/blob/main/src/LanguageServer/Protocol/Handler/Diagnostics/AbstractWorkspacePullDiagnosticsHandler.cs#L94).
So as long as workspace diagnostics requests function this way, it is
easy to get into a situation where there are async listeners that never
complete.

The queue async listener was introduced to ensure that unit tests could
reliably verify if the NFW handler was called (as handling the exception
happened after the response was returned to the client).

To resolve this issue I did a couple things
1. Move reporting NFW for request exceptions into the telemetry
reporting scope. Importantly this scope is called before the result is
sent back to the client. This allows unit tests to simply wait for the
request to complete before checking if the NFW handler was called (no
need for an async listener).
2. Removing the async listener from the queue now that no unit tests
need to wait for an action to occur in the queue.

TODO - validate optprof