Reduce allocations by creating an array builder class that allows span usage
The scrolling speedometer test shows about 1% of CodeAnalysis process allocations due to allocating enumerators in AnalyzerExecutor.ExecuteSyntaxNodeAction. This code currently uses both ArrayBuilder and ImmutableArray, casting down to the common IEnumerable for certain operations. Enumerating then causes an allocation due to not being able to use the struct based enumerator using strongly typed classes allows.

Instead, I've changed the common method to take in a ReadOnlySpan and to enumerate on that. However, that requires the callers to be able to generate a ReadOnlySpan. ImmutableArray can be done via ImmutableCollectionsMarshal, but ArrayBuilder presents a problem as there is no way (that I know of) to get to the underlying ImmutableArray.Builder's array.

To work around this, I've pulled out a minimal subset of code from ArrayBuilder and stripped it's underlying ImmutableArray.Builder and had it hold the array and size directly.