Add runtime async design doc (#75816)
* Add runtime async design doc

Adding an initial design document for runtime async. I'm not making a branch for this yet so I don't have to worry about keeping it up to date until we're actually ready to start implementation work; that should begin soon, but I think getting this document into `main` and starting collaboration with the runtime and debugger folks will help that proceed.

* Remove incorrect IL steps

* Add more examples and updates the runtime document.

* Add another open question

* Add more translated C# samples

* Feedback and more scenarios

* Remove PR link

* Add note about potential hoisted local cleanup