Keep analysis results (including empty) for partial file analysis.
Speedometer scrolling test is showing a huge amount of CPU when moving the caret around a file. It appears that we cache for full compilation and full file. Partial file results were being cached, but not indicating that the analyzers didn't need to be run on that span again if requested.