Reduce allocations from NamedTypeSymbol.AddOperators usage
This method is internal and returned an ImmutableArray. Instead, modify the callers to pass in an ArrayBuilder, as all of them simply enumerate over the returned contents, and don't need to store or create an immutable array from the contents.

This shows up as about 1.5% of total allocations in the codeanalysis process in the scrolling speedometer test.