Fix multiple updates for dynamic file (#76281)
Dynamic file updates can come in multiples before Roslyn asks for them. They are also not guaranteed to not overlap. Rather than trying to reason about a minimum set of updates, this just modifies the response to receive those updates directly. Each update is guaranteed to have no overlapping edits, but across updates that is not guaranteed