Reduce allocations in CompletionItem.GetEntireDisplayText (#76373)
This method's call to string.Concat shows up as about 0.4% of allocations during the typing scenario in the csharp editing speedometer test.

Manual local testing was showing over 90% of calls to this method are done with both prefix and suffix strings empty. Unfortunately, netfx's string.Concat implementation still allocates in this case. Instead, we'll check for this case, and either do the string.Concat if necessary or just use DisplayText directly.