Reduce allocations in ProjectState's ctor (#76369)
* Reduce allocations in ProjectState's ctor

The AsyncLazy allocations in the ProjectState ctor are accounting for 0.6% of allocations during the typing scenario in the csharp editing speedometer test.

The construction of these objects can be deferred until they are queried (or copied). Local testing has about 80-90% of these lazy objects never being used, so this should save at around 0.5% of the allocations in this scenario.