*** WIP ***
Cache the compilation states in compilation order to avoid recalculation

GetStatesInCompilationOrder shows up pretty heavily in CPU usage in the typing scenario in the csharp editing speedometer. The enumerating of the ImmutableList and the repeated searching in the ImmutableSortedDictionary are actually more expensive than one would think.

GetStatesInCompilationOrder is optimized out of the profile, but it looks like it is accounting for about 5% of CPU.

I'm going to run a test insertion with this change and see what the CPU/memory delta of this method looks like with this change.