Optimize ProductDependencyGraph.DoesProjectTransitivelyDependOnProject (#76399)
I believe that this method unnecessarilly obtains the reverse dependencies when the forward references are already defined. If the forward references are present, we should just use the result of trying to find the potential dependencies in it.

DoesProjectTransitivelyDependOnProject shows as 2.9% of allocations and 1.6% of CPU in the typing scenario in the scharp editing speedometer test.