Threading fixes (#76230)
We have a few theoretical races in some of our threading code, as detailed in #75759. I've audited our uses of ImmutableArray as backing fields for this pattern, fixing it up where I find issues, as well as a couple of other threading issues that I found while looking at things named lazy in our codebase.

Closes #75759