Honor abstract required during state initialization in nullable (#76363)
Fixes https://github.com/dotnet/roslyn/issues/74423. When we're getting the list of members to default initialize, we look for properties that override `abstract` properties and have nullable annotations that line up with the base ones. For these properties, we can assume that if we chain to a `SetsRequiredMembers` constructor, it took care of setting a non-null state.