Reduce allocations during checksum creation.
Initially creating this PR as a draft PR to get speedometer results to see if this actually improves things.

From the csharp editing scrolling speedometer, I see about 0.3% of total allocations in the VS process occurring in Checksum.Create. A large part of these allocations are stream/writer constructions. Instead, this PR attempts to use a small pool to reuse and reset these objects.