Release a SeparatedSyntaxListBuilder back to the SyntaxListPool
I noticed this when looking at the SyntaxListPool and that it's allocated debug info was not reducing back to zero. The early return codepaths earlier in the method are commonly hit avoiding the current code from releasing "variables" back into the pool. Changed to account for the early returns by utilizing the existing finally block similar to what is done for the "mods" pooled item.