Reduce allocations in the C# lexer ctor (#76544)
* Attempt to reduce allocations in the C# lexer ctor

The lexer ctor shows as about 0.5% of all allocations over the lifetime of the RoslynCodeAnalysisProcess in the csharp editing speedometer tests.

There was already the concept of the LexerCache which had some pooling usage in it. This PR moves a couple other members to it to allow for their pooling.