Switch ObjectPool to use Interlocked.Exchange instead of CompareExchange
This may be beneficial if the location being queried has had a new non-null value placed into it since the optimistic read, allowing use of that value, instead of skipping it.

When opening Roslyn.sln, I verified both modified Allocate methods did indeed hit the case where Interlocked.Exchange returned a non-value not equal to the optimistic read's value.

I'm a bit hesitant about this change, as the code has not changed for many years and they maybe there is some low-level optimization that such a change will negate. However, on the surface, it seems like using Interlocked.Exchange should be sufficient, so I'll go ahead and create the PR and see if someone else can find something I'm missing.