ResultProvider: Consume DkmClrModuleInstance.GetMetaDataImportHolder (#76395)
Previous to this change, the ResultProvider would use `DkmClrModuleInstance.GetMetaDataImport`, but did NOT use Marshal.ReleaseComObject on the result. This could result in locked files if the RCW took a long time to be finalized.

This PR switches to use a new `GetMetaDataImportHolder` API that adds an IDisposable wrapper struct to take care of this cleanup.

This PR requires VS 17.13 Build 35612.30 or newer.