WIP: Add a pool around SyntaxListPool
Not yet ready for review, I'm going to keep this in draft mode until a speedometer run completes against this changes and shows a decent allocation improvement.

Currently, each LanguageParser/DocumentationCommentParser in C# and Parser/Scanner in VB allocate their own SyntaxListPool. Instead, have SyntaxListPool keep a pool of it's instances and require those objects to pull from that pool.

Additionally, this PR modifies the VB parser to ensure that the Context's FreeStatements have been called. The parser's Context manipulation is quite complicated, and before this change about 10% of the BlockContexts created weren't having FreeStatements called on them, and thus not releasing _statements back into the parser's pool. By hooking into the parser's dispose method and having it ensure that FreeStatements is invoked on the last context, this reduces to about 1%. The manipulation of the contexts is too confusing for me to feel comfortable about a more complete solution.