Reduce work in ConversionsBase.AddUserDefinedConversionsToExplicitCandidateSet
The ArrayBuilder usage in this method is showing up slightly in speedometer CPU profiles (0.1%) but more heavily in local profiles I've taken using a much larger version of the file than what the speedometer tests use (2.6%)

This method is called quite often and the (isExplicit && isChecked) check seems to not hit very often (specifically, never in my testing of opening/simple editing of the file). The usage of the pool and populating the array builder aren't necessary unless that case is hit, so I've reshuffled the code around a bit to avoid that.