Start sharing code
Shared contains mostly contracts and a few data classes/wrappers that
are common between VS/VS Code. The layering itself is not strictly for
those clients but instead is meant to be a split to remove the
dependency upon EditorFeatures.

Prepare for Features layer to load an IRazorServiceProvider to be used
in Microsoft.CodeAnalysis.LanguageServer

Remove a lot of the implementation logic from
Microsoft.CodeAnalysis.LanguageServer. This should now be provided from
the razor assembly that is loaded in the LanguageServer via
RazorLSPServiceProvider

Remove shared code from the existing EA.