Move internal tools restore out of CI yaml and into eng/build.ps1 (#76911)
* Move internal tools restore out of CI yaml and into eng/build.ps1
The internal tools must be restored before loading the powershell module to find IBCM merge data. To work nicely with VMR builds, which call roslyn's eng/build.ps1 script, move this restore logic out of the YAML and into the eng/build.ps1 script. The tools are only restored when we know we need to load the module. Note that the tools project will already be restored again automatically when roslyn builds because of the use of IBC data.

* Update arcade and fixup dependencies