Create up-to-date check input items for the publish inputs
Without this change, it's possible to make a change in the regular
product but not get ServiceHub to rebuild. This is because the output
group we consume in the VSIX consumes the published output (by
necessity) but doesn't trigger the publish itself. The publish is
triggered by the PublishProjectReferencesForVsixCreation in the same
file prior to the VSIX build. But if the fast-up-to-date check sees
the project as up to date (because no publish happened), then we'll
never actually run the publish in the first place.

The solution here is to add items that get published to the list of
VSIX inputs, so that way we'll trigger the full build (and publish)
correctly.