Change override completion to select text after updating the buffer.
This (along with https://github.com/dotnet/roslyn/pull/76969) will allow a user to commit an override completion and easily replace the default text (and still use the expression body fixers).

This required the CompletionChange class to now allow the data to represent a selection instead of just a cursor position.