Realize less of the syntax tree during AbstractSemanticModelReuseLanguageService.GetPreviousBodyNode (#77032)
Realize less of the syntax tree during AbstractSemanticModelReuseLanguageService.GetPreviousBodyNode

During the typing scenario in the scrolling speedometer test, about 13% of allocations are under GetPreviousBodyNode, due to realizing a bunch of nodes in the syntax tree to attempt to map from a method node in one tree to the corresponding node in similarly structured tree. Instead, as we know the trees have similar structure, just walk up the ancestors in the current tree to generate a node index path, and apply that path to the previous tree to find the appropriate method node.