Set a RID for InteractiveHost32 (#77116)
The .NET SDK's RID (and PlatformTarget) inference for .NET Framework only executes on Windows. As a result, building InteractiveHost32 on non-Windows produces a differently-named NuGet package. This is causing us issues in the VMR.

Set the RID explicitly like the InteractiveHost64 process does. This ensures that the resulting package name is InteractiveHost32.x86.Symbols when built on Windows or non-Windows.