Add new workspace event that gives handlers the opportunity to be processed immediately (#76932)
Add new workspace event that gives handlers the opportunity to be processed immediately

This is in response to allocations seen when the OOP process attempts to operate on a sourcetext version that has not been synchronized over. In those cases, the OOP process requests a full deserialization of the source text causing full buffer allocations in both VS and the CA process.

There are several points of asynchronicity in the current system around sending over buffer changes, reducing any of which would reduce the likelihood of needing this full serialization.

This PR remove one of point of asynchronicity in the workspace eventing layer and reduces another point of asynchronicity by bypassing the old _textChangeQueue ABWQ and doing a fire and forget instead. Previously, all notifications were done on a delayed basis, by wrapping each notification in the WorkspaceSpace.ScheduleTask call from the RaiseWorkspaceChangedEventAsync method. This method allows callers to indicate they need to be called immediately upon the workspace change. As noted by a doc comment in the PR, these handlers should be very fast.