Update to release nuget packages (#77261)
* Update to release nuget packages

In order to publish 17.13 nuget packages, we need to update a few dependencies. I updated the minimum number of packages required from our public packages, then fixed up any restore issues. For each updated package, I then went and confirmed that the dll was actually present in 17.13.0, so we should be safe from dependency loading issues. This change does not need to be inserted into VS.

* Skip tests for https://github.com/microsoft/vs-solutionpersistence/issues/95

* Skip failing tests, see https://github.com/dotnet/roslyn/issues/77293.