Null annotate MetadataHelpers
Issue #76394 indicites there is a null reference happening inside
`MetadataHelpers.GetInfoForImmediateNamespaceMembers`. The symptoms
indicate it's non-deterministic (deterministic ordering or race
condition).

Decided to null annotate the type and a few surrounding calls to see if
I could get lucky and spot the source of the NRE. Was not able to do
this but decided to submit the annotations.