Don't realize the SourceText in SyntaxTree.OverlapsHiddenPosition if not needed
This shows up in override completion commit in the speedometer profiles. Going to start in draft mode and start a speedometer run to see if this helps, or somewhere else just realizes the SourceText during commit if not done here.

Will update with more perf numbers if the speedometer run looks good.