Fix issue in watch window completion (#77204)
* Fix issue in watch window completion

The issue here is when the watch window contains

for (int i = 0; i < 10; i++)
    [Console.WriteLine(i);]

where [] denotes the current statement. In that case, the adjusted context should be at the start of Console.WriteLine. Additionally, a semicolon should not be placed in the generated projection buffer before debuggerMappedSpan, as that would place debuggerMappedSpan outside the for loop.

Fixes https://github.com/dotnet/roslyn/issues/42718