Fix design time build error in Roslyn.VisualStudio.Setup
The MSBuild task invocation here is a copy/paste from earlier in the
file, but the item group renamed so it didn't conflict. The rename
was in complete, so the use in the Properties property was still
using the old name; this had the effect that we wouldn't set the TFM
property, which would cause problems for multi-targeting projects.